+++ lib/no_namespace/focuspoint_media.php
@@ -16,9 +16,10 @@
  * auf diesem Wege weiterhin bereitgestellt. Diese Klasse fällt mit Version 5.0.0 weg, nachdem
  * Redaxo auf Composer als Installer umgestellt wurde.
  *
- * @deprecated 5.0.0 Aufrufe auf "FriendsOfREDAXO\focuspoint\focuspoint_media" (Namespace) umstellen
+ * @deprecated 5.0.0 Aufrufe auf "FriendsOfRedaxo\focuspoint\focuspoint_media" (Namespace) umstellen
+ * @see https://github.com/orgs/FriendsOfREDAXO/discussions/40
  */
 
-class focuspoint_media extends FriendsOfREDAXO\focuspoint\focuspoint_media
+class focuspoint_media extends FriendsOfRedaxo\focuspoint\focuspoint_media
 {
 }