+++ uninstall.php
@@ -26,7 +26,7 @@
  *  @var rex_addon $this
  */
 
-namespace FriendsOfREDAXO\focuspoint;
+namespace FriendsOfRedaxo\focuspoint;
 
 use rex;
 use rex_effect_abstract_focuspoint;